Fix `LICENSE` file using GPL instead of LGPL:
This project is supposed to be licensed under the LGPL since it is a package,
and licensing under the GPL would mean anyone using this package would have to
license their project under a GPL license as well, which is not what I want.
The `package.json` correctly uses the LGPL license, but when I copied the
`LICENSE` file I copied the GPL instead of the LGPL. This has now been fixed.